Data.Expression

Description

Usage:

You can build expressions in predefined languages (QFLogic, QFLia, QFALia, Lia, ALia) using the smart constructors such as var, and, or, not, forall, exists (or operators .&., .|., .->., .<-., .<->.) or you can define your own sorted language as a fixpoint (IFix) of a sum (:+:) of indexed functors (IFunctor).

Documentation

module Data.Expression.Arithmetic

module Data.Expression.Array

module Data.Expression.Equality

module Data.Expression.IfThenElse

module Data.Expression.Parser

data family Sing (a :: k) :: Type

The singleton kind-indexed data family.

data Sort

A universe of recognized sorts

Constructors

BooleanSort
booleans (true, false)

IntegralSort
integers (..., -1, 0, 1, ...)

ArraySort Sort Sort
arrays indexed by Sort sort, containing elements of element sort

type DynamicallySortedFix f = DynamicallySorted (IFix f)

An expression of some sort (obtained for example during parsing)

data DynamicallySorted (d :: Sort -> *) where

A value of some sort

Constructors

DynamicallySorted :: forall (s :: Sort) d. Sing s -> d s -> DynamicallySorted d

data DynamicSort where

Some sort (obtained for example during parsing)

Constructors

DynamicSort :: forall (s :: Sort). Sing s -> DynamicSort

withSort :: forall a (s :: Sort). Sing s -> (SingI s => a) -> a

Turn implicit sort parameter into explicit one

toStaticSort :: forall (s :: Sort). SingI s => DynamicSort -> Maybe (Sing s)

Tries to convert some sort (DynamicSort) to a requested sort.

toDynamicallySorted :: forall d (s :: Sort). SingI s => d s -> DynamicallySorted d

Converts a statically sorted expression to a dynamically sorted one.

toStaticallySorted :: forall d (s :: Sort). SingI s => DynamicallySorted d -> Maybe (d s)

Tries to convert an expression (DynamicallySorted) of some sort to an expression of requested sort. Performs no conversions.

parseSort :: Parser DynamicSort

Parser that accepts sort definitions such as bool, int, array int int, array int (array ...).

module Data.Expression.Utils.Indexed.Eq

module Data.Expression.Utils.Indexed.Foldable

module Data.Expression.Utils.Indexed.Functor

module Data.Expression.Utils.Indexed.Show

module Data.Expression.Utils.Indexed.Sum

module Data.Expression.Utils.Indexed.Traversable

type QFLogicF = EqualityF :+: (ConjunctionF :+: (DisjunctionF :+: (NegationF :+: VarF)))

A functor representing propositional logic embedded in first order logic (quantifier-free, boolean variables aka propositions, logical connectives and, or, not, equality of propositions)

type QFLiaF = ArithmeticF :+: (IfThenElseF :+: QFLogicF)

A functor representing the language of quantifier-free linear integer arithmetic theory in first order logic (integer constants, integer variables, addition, multiplication, divisibility, ordering)

type LiaF = ExistentialF BooleanSort :+: (ExistentialF IntegralSort :+: (UniversalF BooleanSort :+: (UniversalF IntegralSort :+: QFLiaF)))

A functor much like QFLiaF with quantifiers over booleans and integers

type QFALiaF = ArrayF :+: QFLiaF

A functor representing the language of quantifier-free linear integer arithmetic and array theories in first order logic (much like QFLiaF with additional array variables, select, and store)

type ALiaF = ExistentialF BooleanSort :+: (ExistentialF IntegralSort :+: (UniversalF BooleanSort :+: (UniversalF IntegralSort :+: QFALiaF)))

A functor much like QFALiaF with quantifiers over booleans and integers

type Var = IFix VarF

A language consisting solely of variables (useful for listing variables outside of any particular context, such as bound variables of quantified formula)

type QFLogic = IFix QFLogicF

A language obtained as fixpoint of QFLogicF

type QFLia = IFix QFLiaF

A language obtained as fixpoint of QFLiaF

type Lia = IFix LiaF

A language obtained as fixpoint of LiaF

type QFALia = IFix QFALiaF

A language obtained as fixpoint of QFALiaF

type ALia = IFix ALiaF

A language obtained as fixpoint of ALiaF

class BoundedLattice a => ComplementedLattice a where

Bounded lattices that support complementing elements

complement . complement = id

Methods

complement :: a -> a

type VariableName = String

Type of names assigned to variables

data VarF a (s :: Sort) where

A functor representing a sorted variable, each variable is identified by its name and sort

Constructors

Var :: VariableName -> Sing s -> VarF a s

data ConjunctionF a (s :: Sort) where

A functor representing a logical connective for conjunction

Constructors

And :: [a BooleanSort] -> ConjunctionF a BooleanSort

data DisjunctionF a (s :: Sort) where

A functor representing a logical connective for disjunction

Constructors

Or :: [a BooleanSort] -> DisjunctionF a BooleanSort

data NegationF a (s :: Sort) where

A functor representing a logical connective for negation

Constructors

Not :: a BooleanSort -> NegationF a BooleanSort

data UniversalF (v :: Sort) a (s :: Sort) where

A functor representing a mono-sorted universal quantifier binding a number of variables within a formula

Constructors

Forall :: [Var v] -> a BooleanSort -> UniversalF v a BooleanSort

data ExistentialF (v :: Sort) a (s :: Sort) where

A functor representing a mono-sorted existential quantifier binding a number of variables within a formula

Constructors

Exists :: [Var v] -> a BooleanSort -> ExistentialF v a BooleanSort

newtype Substitution f

Substitution that given an expression produces replacement if the expression is to be replaced or nothing otherwise.

Constructors

Substitution

Fields

runSubstitution :: forall (s :: Sort). IFix f s -> Maybe (IFix f s)

substitute :: (IFunctor f, IEq1 f) => IFix f s -> Substitution f -> IFix f s

Executes a substitution.

for :: forall f (s :: Sort). (IFunctor f, IEq1 f) => IFix f s -> IFix f s -> Substitution f

A simple constructor of substitutions that replaces the latter expression with the former.

var :: forall f s. (VarF :<: f, SingI s) => VariableName -> IFix f s

A smart constructor for variables of any sort in any language Takes the variable name and infers the target language and sort from context.

var "a" :: Lia 'IntegralSort

dynvar :: forall f (s :: Sort). VarF :<: f => VariableName -> Sing s -> DynamicallySortedFix f

Like var except it hides the sort inside DynamicallySortedFix

true :: ConjunctionF :<: f => IFix f BooleanSort

Logical tautology

false :: DisjunctionF :<: f => IFix f BooleanSort

Logical contradiction

and :: ConjunctionF :<: f => [IFix f BooleanSort] -> IFix f BooleanSort

A smart constructor for variadic conjunction

or :: DisjunctionF :<: f => [IFix f BooleanSort] -> IFix f BooleanSort

A smart constructor for variadic disjunction

not :: NegationF :<: f => IFix f BooleanSort -> IFix f BooleanSort

A smart constructor for negation

forall :: UniversalF v :<: f => [Var v] -> IFix f BooleanSort -> IFix f BooleanSort

A smart constructor for universally quantified formulae

exists :: ExistentialF v :<: f => [Var v] -> IFix f BooleanSort -> IFix f BooleanSort

A smart constructor for existentially quantified formulae

(.&.) :: ConjunctionF :<: f => IFix f BooleanSort -> IFix f BooleanSort -> IFix f BooleanSort infixr 6

A smart constructor for binary conjunction

(.|.) :: DisjunctionF :<: f => IFix f BooleanSort -> IFix f BooleanSort -> IFix f BooleanSort infixr 5

A smart constructor for binary disjunction

(.->.) :: (DisjunctionF :<: f, NegationF :<: f) => IFix f BooleanSort -> IFix f BooleanSort -> IFix f BooleanSort infixr 4

A smart constructor for implication (an abbreviation for not a .|. b)

(.<-.) :: (DisjunctionF :<: f, NegationF :<: f) => IFix f BooleanSort -> IFix f BooleanSort -> IFix f BooleanSort infixl 4

A smart constructor for reversed implication (an abbreviation for a .|. not b)

(.<->.) :: (ConjunctionF :<: f, DisjunctionF :<: f, NegationF :<: f) => IFix f BooleanSort -> IFix f BooleanSort -> IFix f BooleanSort infix 3

A smart constructor for if-and-only-if connective

(./=.) :: forall f s. (NegationF :<: f, EqualityF :<: f, SingI s) => IFix f s -> IFix f s -> IFix f BooleanSort infix 7

A smart constructor for disequality

literals :: (ConjunctionF :<: f, DisjunctionF :<: f) => IFix f BooleanSort -> [IFix f BooleanSort]

literals decomposes a boolean combination (formed with conjunctions and disjunctions, preferably in negation normal form) into its constituents.

conjuncts :: ConjunctionF :<: f => IFix f BooleanSort -> [IFix f BooleanSort]

conjuncts decomposes a conjunction into conjuncts.

disjuncts :: DisjunctionF :<: f => IFix f BooleanSort -> [IFix f BooleanSort]

disjuncts decomposes a disjunction into disjuncts.

vars :: (VarF :<: f, IFoldable f, IFunctor f) => IFix f s -> [DynamicallySorted Var]

Collects a list of all variables occurring in an expression (bound or free).

freevars :: (IFunctor f, MaybeQuantified f) => IFix f s -> [DynamicallySorted Var]

Collects a list of all free variables occurring in an expression.

freenames :: forall f (s :: Sort). (VarF :<: f, IFunctor f, IFoldable f) => IFix f s -> VariableNamePool

class MaybeQuantified f

Minimal complete definition

isQuantified', freevars'

isQuantified :: MaybeQuantified f => IFix f s -> Bool

Test whether an expression contains a quantifier.

isQuantifierFree :: MaybeQuantified f => IFix f s -> Bool

Tests whether an expression is free of any quantifier.

class (NegationF :<: f, HasDual f f) => NNF f

nnf :: forall f. NNF f => IFix f BooleanSort -> IFix f BooleanSort

Propagates negation toward boolean atoms (across conjunction, disjunction, quantifiers).

class (VarF :<: f, NegationF :<: f, IFunctor f, IFoldable f, ITraversable f, HasDual f f, MaybeQuantified' f f) => Prenex f

prenex :: forall f. Prenex f => IFix f BooleanSort -> IFix f BooleanSort

Puts an expression into prenex form (quantifier prefix and a quantifier-free formula).

class (VarF :<: f, Bind f f, Bind' f f, MaybeQuantified'' f f, IFoldable f, ITraversable f) => Flatten f

flatten :: forall f. Flatten f => IFix f BooleanSort -> IFix f BooleanSort

Replaces non-variable and non-constant arguments to uninterpreted functions (such as select and store) with a fresh bound (universally or existentially) variable that is bound to the original term.

class (VarF :<: f, EqualityF :<: f, Bind f f, Bind' f f, MaybeQuantified'' f f, Forall f f, Axiomatized f f, IFoldable f, ITraversable f) => Unstore f

unstore :: forall f. Unstore f => IFix f BooleanSort -> IFix f BooleanSort

Replaces store with an instance of its axiomatization.