Data.Expression.Array

Description

Documentation

data ArrayF a (s :: Sort) where

A functor representing array-theoretic terms (select and store also known as "read" and "write")

Constructors

Select :: Sing i -> Sing e -> a (ArraySort i e) -> a i -> ArrayF a e

Store :: Sing i -> Sing e -> a (ArraySort i e) -> a i -> a e -> ArrayF a (ArraySort i e)

select :: (ArrayF :<: f, SingI i, SingI e) => IFix f (ArraySort i e) -> IFix f i -> IFix f e

A smart constructor for select

store :: (ArrayF :<: f, SingI i, SingI e) => IFix f (ArraySort i e) -> IFix f i -> IFix f e -> IFix f (ArraySort i e)

A smart constructor for store

accesses :: forall f (s :: Sort). (ArrayF :<: f, ITraversable f) => IFix f s -> [DynamicArrayAccess f]

Collects pairs of arrays and indices that appear together in some select and/or store within an expression

toStaticallySortedArrayAccess :: forall f (i :: Sort) (e :: Sort). (SingI i, SingI e) => DynamicArrayAccess f -> Maybe (IFix f (ArraySort i e), IFix f i)

Tries to convert access to an array of some sort to an access to an array of a specific sort

newtype ArrayAccess f (i :: Sort) (e :: Sort)

Constructors

ArrayAccess

Fields

getAA :: (Array f e i, Index f i)

newtype DynamicValueArrayAccess f (i :: Sort)

Constructors

DynamicValueArrayAccess

Fields

getDVAA :: DynamicallySorted (ArrayAccess f i)

type DynamicArrayAccess f = DynamicallySorted (DynamicValueArrayAccess f)