Data.Expression.Parser

Description

Documentation

type Context = Map String DynamicSort

Parsing context assigning sorts to known variables

type Parser = ReaderT Context (StateT Context Parser)

Context-sensitive parser that remembers sorts of variables

class Parseable f g where

Expressions that can be parsed

Methods

parser :: Proxy f -> Parser (DynamicallySortedFix g) -> Parser (DynamicallySortedFix g)

parse :: forall f (s :: Sort). (Parseable f f, SingI s) => Text -> Maybe (IFix f s)

Tries to parse the entire input text and produce an expression in desired language and with desired sort.

parseWith :: forall f (s :: Sort). (Parseable f f, SingI s) => Context -> Text -> Maybe (IFix f s)

Like parse but allows adding initial assumption about sorts of some variables.

char :: Char -> Parser Char

Matches a given character.

choice :: [Parser a] -> Parser a

Matches first of many choices.

decimal :: Parser Int

Matches a given decimal number.

digit :: Parser Char

Matches a given digit.

letter :: Parser Char

Matches any character.

many1 :: Parser a -> Parser [a]

Matches one or more times.

sepBy1 :: Parser a -> Parser s -> Parser [a]

Matches one or more times, separated by specified separator.

signed :: Parser Int -> Parser Int

Matches a signed number.

space :: Parser Char

Matches space.

string :: Text -> Parser Text

Matches a given string.

identifier :: Parser String

Matches identifier that starts with [a-zA-Z'_#] and is followed by [a-zA-Z'_#0-9].

(<?>) :: Parser a -> String -> Parser a

Labels parser.

assertSort :: String -> DynamicSort -> Parser ()

Asserts what sort is assigned to a variable in current context.

assumeSort :: String -> Parser DynamicSort

Variable assumes sort based on current context.