Data.Expression.Sort

Description

Documentation

data Sort

A universe of recognized sorts

Constructors

BooleanSort
booleans (true, false)

IntegralSort
integers (..., -1, 0, 1, ...)

ArraySort
arrays indexed by Sort sort, containing elements of element sort

Fields

index :: Sort

element :: Sort

data family Sing (a :: k) :: Type

The singleton kind-indexed data family.

withSort :: forall a (s :: Sort). Sing s -> (SingI s => a) -> a

Turn implicit sort parameter into explicit one

data DynamicSort where

Some sort (obtained for example during parsing)

Constructors

DynamicSort :: forall (s :: Sort). Sing s -> DynamicSort

data DynamicallySorted (d :: Sort -> *) where

A value of some sort

Constructors

DynamicallySorted :: forall (s :: Sort) d. Sing s -> d s -> DynamicallySorted d

type DynamicallySortedFix f = DynamicallySorted (IFix f)

An expression of some sort (obtained for example during parsing)

parseSort :: Parser DynamicSort

Parser that accepts sort definitions such as bool, int, array int int, array int (array ...).

toDynamicallySorted :: forall d (s :: Sort). SingI s => d s -> DynamicallySorted d

Converts a statically sorted expression to a dynamically sorted one.

toStaticSort :: forall (s :: Sort). SingI s => DynamicSort -> Maybe (Sing s)

Tries to convert some sort (DynamicSort) to a requested sort.

toStaticallySorted :: forall d (s :: Sort). SingI s => DynamicallySorted d -> Maybe (d s)

Tries to convert an expression (DynamicallySorted) of some sort to an expression of requested sort. Performs no conversions.